backend/pyproject.toml
[project]
name = "urgencyflow-backend"
version = "0.1.0"
description = "Emergency room AI priority queue system - Backend API"
requires-python = ">=3.11"
dependencies = [
    # Django
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.5.0",

    # Machine Learning
    "torch>=2.5.0",
    "pandas>=2.2.3",
    "numpy>=2.1.2",
    "scikit-learn>=1.5.2",
    "imbalanced-learn>=0.12.4",
    "joblib>=1.4.2",

    # Visualization
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",
]

[tool.uv]
dev-dependencies = [
    "ipython>=8.28.0",
    "ipykernel>=6.29.5",
]
